fix: prevent floats without decimal values from being encoded as int
Floats are encoded as ints and trailing zeroes are being stripped,
causing document rejection due to mismatched types.
APM server is relying on fastjson, which is using strconv, to encode
metrics.
The precision is set to -1 so that the encoded string uses the
smallest number of digits. To avoid losing type information, truncate
the value by casting to int and compare the result to the original
value: if they are equal conversion is carried out with precision=1
so that the trailing zero is not stripped.

+++ systemtest/metrics_test.go
@@ -170,7 +170,7 @@ func TestApplicationMetrics(t *testing.T) {
 			"full_name": "a.b.c",
 			"mapping": map[string]interface{}{
 				"c": map[string]interface{}{
-					"type": "long",
+					"type": "float",
 				},
 			},
 		},
